chore: - Configured project with Firebase Firestore SDK Libraries.        - Implemented Staff Registration to Add New Lecturers to Firestore Db
This commit completes Firebase or Database related work for Staff/Lecturer registration. But it does not implement Student Registration. This is because Student Registration has some special requirements I would like to run by the team before implementing it.

The implementation of Staff Registration adds 3 more files: Student.java, Course.java and Lecturer.java. These files are representation entity files that are used as templates to save Lecturers and Students (when implemented) in the Firestore Database